Grype facade string sbom (#988)
* Pass sbom as a string or as a file. Add a method to return the body of metadata.json. Protect against archive file that doesn't exist Add more tests. Refactor state variables to minimize the use of global.

+++ anchore_engine/clients/grype_wrapper.py
@@ -1,6 +1,7 @@
 import anchore_engine.configuration.localconfig
-import os
+import errno
 import json
+import os
 import shlex
 import shutil
 import sqlalchemy
@@ -8,31 +9,26 @@ import tarfile
 
 from anchore_engine.db.entities.common import UtilMixin
 from anchore_engine.subsys import logger
-from anchore_engine.utils import run_check, CommandException
+from anchore_engine.utils import CommandException, run_check, run_piped_command_list
+from json.decoder import JSONDecodeError
 from readerwriterlock import rwlock
 from sqlalchemy import Column, String, Integer, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
 
-grype_db_dir = None
-grype_db_session = None
+__grype_db_dir = None
+__grype_db_session = None
 
 grype_db_lock = rwlock.RWLockWrite()
 Base = declarative_base()
 
+GRYPE_SUB_CMD = "grype -vv -o json"
 VULNERABILITY_FILE_NAME = "vulnerability.db"
+METADATA_FILE_NAME = "metadata.json"
 VULNERABILITY_TABLE_NAME = "vulnerability"
 VULNERABILITY_METADATA_TABLE_NAME = "vulnerability_metadata"
 
-
-# TODO Implement, once added to Grype
-def get_current_grype_db_checksum():
-    """
-    Return the checksum for the in-use version of grype db
-    """
-    grype_db_checksum = None
-    logger.info("Returning current grype_db checksum: {}".format(grype_db_checksum))
-    return grype_db_checksum
+ARCHIVE_FILE_NOT_FOUND_ERROR_MESSAGE = "New grype_db archive file not found"
 
 
 def _get_default_grype_db_dir_from_config():
@@ -41,16 +37,16 @@ def _get_default_grype_db_dir_from_config():
     """
     localconfig = anchore_engine.configuration.localconfig.get_config()
     if "grype_db_dir" in localconfig:
-        grype_db_dir = os.path.join(
+        local_grype_db_dir = os.path.join(
             localconfig["service_dir"], localconfig["grype_db_dir"]
         )
     else:
-        grype_db_dir = os.path.join(localconfig["service_dir"], "grype_db/")
+        local_grype_db_dir = os.path.join(localconfig["service_dir"], "grype_db/")
 
-    if not os.path.exists(grype_db_dir):
-        os.mkdir(grype_db_dir)
+    if not os.path.exists(local_grype_db_dir):
+        os.mkdir(local_grype_db_dir)
 
-    return grype_db_dir
+    return local_grype_db_dir
 
 
 def _move_grype_db_archive(
@@ -60,16 +56,30 @@ def _move_grype_db_archive(
     archive_file_name = os.path.basename(grype_db_archive_local_file_location)
     grype_db_archive_copied_file_location = os.path.join(output_dir, archive_file_name)
 
-    # Move the archive file
-    logger.info(
-        "Moving the grype_db archive from {} to {}".format(
+    if not os.path.exists(grype_db_archive_local_file_location):
+        logger.warn(
+            "Unable to move grype_db archive from {} to {} because it does not exist".format(
+                grype_db_archive_local_file_location,
+                grype_db_archive_copied_file_location,
+            )
+        )
+        raise FileNotFoundError(
+            errno.ENOENT,
+            ARCHIVE_FILE_NOT_FOUND_ERROR_MESSAGE,
+            grype_db_archive_local_file_location,
+        )
+    else:
+        # Move the archive file
+        logger.info(
+            "Moving the grype_db archive from {} to {}".format(
+                grype_db_archive_local_file_location,
+                grype_db_archive_copied_file_location,
+            )
+        )
+        os.replace(
             grype_db_archive_local_file_location, grype_db_archive_copied_file_location
         )
-    )
-    os.replace(
-        grype_db_archive_local_file_location, grype_db_archive_copied_file_location
-    )
-    return grype_db_archive_copied_file_location
+        return grype_db_archive_copied_file_location
 
 
 def _open_grype_db_archive(
@@ -114,7 +124,7 @@ def _move_and_open_grype_db_archive(
     # Get the location to copy the archive to
     local_db_dir = _get_default_grype_db_dir_from_config()
 
-    # Copy the archive
+    # Move the archive
     grype_db_archive_copied_file_location = _move_grype_db_archive(
         grype_db_archive_local_file_location, local_db_dir
     )
@@ -131,7 +141,7 @@ def _move_and_open_grype_db_archive(
     return latest_grype_db_dir
 
 
-def _init_grype_db_engine(latest_grype_db_dir):
+def _init_latest_grype_db_engine(latest_grype_db_dir):
     """
     Create and return the sqlalchemy engine object
     """
@@ -144,7 +154,7 @@ def _init_grype_db_engine(latest_grype_db_dir):
     return latest_grype_db_engine
 
 
-def _init_grype_db_session(grype_db_engine):
+def _init_latest_grype_db_session(grype_db_engine):
     """
     Create and return the db session
     """
@@ -158,22 +168,16 @@ def _init_grype_db_session(grype_db_engine):
     return grype_db_session
 
 
-def _init_grype_db(lastest_grype_db_archive: str, version_name: str):
+def _init_latest_grype_db(lastest_grype_db_archive: str, version_name: str):
     """
     Write the db string to file, create the engine, and create the session
     Return the file and session
     """
-    # TODO Further refine this, we should be extracting everything to a subdir, and letting the
-    # replacement logic operate at that level (ie always write and remove the contents of an
-    # archive file to a subdir under '<service-dir>/grype_db/', rather than operating on
-    # the specific vulnerability.db file.
-    # It sounds like we will also be getting the metata.json file, and the further contents
-    # of those archives could change if they needed to.
     latest_grype_db_dir = _move_and_open_grype_db_archive(
         lastest_grype_db_archive, version_name
     )
-    latest_grype_db_engine = _init_grype_db_engine(latest_grype_db_dir)
-    latest_grype_db_session = _init_grype_db_session(latest_grype_db_engine)
+    latest_grype_db_engine = _init_latest_grype_db_engine(latest_grype_db_dir)
+    latest_grype_db_session = _init_latest_grype_db_session(latest_grype_db_engine)
 
     # Return the dir, file, and engine
     return latest_grype_db_dir, latest_grype_db_session
@@ -195,12 +199,13 @@ def _remove_local_grype_db(grype_db_dir):
     return
 
 
-def update_grype_db(grype_db_archive_local_file_location: str, version_name: str):
+def __init_grype_db_engine(
+    grype_db_archive_local_file_location: str, version_name: str
+):
     """
     Update the installed grype db with the provided definition, and remove the old grype db file.
     This method does not validation of the db, and assumes it has passed any required validation upstream
     """
-    global grype_db_dir, grype_db_session
 
     logger.info(
         "Updating grype with a new grype_db archive from {}".format(
@@ -214,15 +219,15 @@ def update_grype_db(grype_db_archive_local_file_location: str, version_name: str
 
             # Store the db locally and
             # Create the sqlalchemy engine for the new db
-            latest_grype_db_dir, latest_grype_db_session = _init_grype_db(
+            latest_grype_db_dir, latest_grype_db_session = _init_latest_grype_db(
                 grype_db_archive_local_file_location, version_name
             )
 
             # Store the dir and session variables globally
             # For use during reads and to remove in the next update
-            old_grype_db_dir = grype_db_dir
-            grype_db_dir = latest_grype_db_dir
-            grype_db_session = latest_grype_db_session
+            old_grype_db_dir = _get_grype_db_dir()
+            _set_grype_db_dir(latest_grype_db_dir)
+            _set_grype_db_session(latest_grype_db_session)
 
             # Remove the old local db
             if old_grype_db_dir:
@@ -231,32 +236,119 @@ def update_grype_db(grype_db_archive_local_file_location: str, version_name: str
             write_lock.release()
 
 
-def get_vulnerabilities(grype_sbom: str) -> json:
+def _get_grype_db_dir():
+    return __grype_db_dir
+
+
+def _get_grype_db_session():
+    return __grype_db_session
+
+
+def _set_grype_db_dir(grype_db_dir):
+    global __grype_db_dir
+    __grype_db_dir = grype_db_dir
+
+
+def _set_grype_db_session(grype_db_session):
+    global __grype_db_session
+    __grype_db_session = grype_db_session
+
+
+def get_current_grype_db_metadata() -> json:
+    """
+    Return the json contents of the metadata file for the in-use version of grype db
+    """
+    # Get the path to the latest grype_db metadata file
+    latest_grype_db_metadata_file = os.path.join(
+        _get_grype_db_dir(), METADATA_FILE_NAME
+    )
+
+    # Ensure the file exists
+    if not os.path.exists(latest_grype_db_metadata_file):
+        # If not, return None
+        return None
+    else:
+        # Get the contents of the file
+        with open(latest_grype_db_metadata_file) as read_file:
+            try:
+                return json.load(read_file)
+            except JSONDecodeError:
+                logger.error(
+                    "Unable to decode grype_db metadata file into json: %s", read_file
+                )
+                return None
+
+
+def _get_proc_env():
+    # Set grype env variables, including the grype db location
+    grype_env = {
+        "GRYPE_CHECK_FOR_APP_UPDATE": "0",
+        "GRYPE_LOG_STRUCTURED": "1",
+        "GRYPE_DB_AUTO_UPDATE": "0",
+        "GRYPE_DB_CACHE_DIR": "{}".format(_get_grype_db_dir()),
+    }
+    proc_env = os.environ.copy()
+    proc_env.update(grype_env)
+    return proc_env
+
+
+def get_vulnerabilities_for_sbom(grype_sbom: str) -> json:
     """
     Use grype to scan the provided sbom for vulnerabilites.
     """
-    global grype_db_dir
-
     # Get the read lock
     read_lock = grype_db_lock.gen_rlock()
     if read_lock.acquire(blocking=False, timeout=60):
         try:
-            # Set grype env variables, including the grype db location
-            grype_env = {
-                "GRYPE_CHECK_FOR_APP_UPDATE": "0",
-                "GRYPE_LOG_STRUCTURED": "1",
-                "GRYPE_DB_AUTO_UPDATE": "0",
-                "GRYPE_DB_CACHE_DIR": "{}".format(grype_db_dir),
-            }
+            # Get env variables to run the grype scan with
+            proc_env = _get_proc_env()
 
-            proc_env = os.environ.copy()
-            proc_env.update(grype_env)
+            # Format and run the command. Grype supports piping in an sbom string, so we need to this in two steps.
+            # 1) Echo the sbom string to std_out
+            # 2) Pipe that into grype
+            pipe_sub_cmd = "echo '{sbom}'".format(
+                sbom=grype_sbom,
+            )
+            full_cmd = [shlex.split(pipe_sub_cmd), shlex.split(GRYPE_SUB_CMD)]
+
+            logger.debug(
+                "Running grype with command: {} | {}".format(
+                    pipe_sub_cmd, GRYPE_SUB_CMD
+                )
+            )
+
+            stdout = None
+            err = None
+            try:
+                _, stdout, _ = run_piped_command_list(full_cmd, env=proc_env)
+            except CommandException as exc:
+                logger.error(
+                    "Exception running command: {} | {}, stderr: {}".format(
+                        pipe_sub_cmd, GRYPE_SUB_CMD, exc.stderr
+                    )
+                )
+                raise exc
+        finally:
+            read_lock.release()
+
+        # Return the output as json
+        return json.loads(stdout.decode("utf-8"))
+
+
+def get_vulnerabilities_for_sbom_file(grype_sbom_file: str) -> json:
+    """
+    Use grype to scan the provided sbom for vulnerabilites.
+    """
+    # Get the read lock
+    read_lock = grype_db_lock.gen_rlock()
+    if read_lock.acquire(blocking=False, timeout=60):
+        try:
+            # Get env variables to run the grype scan with
+            proc_env = _get_proc_env()
 
             # Format and run the command
-            # TODO This is currently expecting an sbom file reference, not the actual sbom string itself
-            # Should we support both, or just the string?
-            cmd = "grype -vv -o json sbom:{sbom}".format(
-                sbom=grype_sbom,
+            cmd = "{grype_sub_command} sbom:{sbom}".format(
+                grype_sub_command=GRYPE_SUB_CMD, sbom=grype_sbom_file
             )
 
             logger.debug("Running grype with command: {}".format(cmd))
@@ -316,8 +408,6 @@ def query_vulnerabilities(
     Query the grype db for vulnerabilites. affected_package_version is unused, but is left in place for now to match the
     header of the existing function this is meant to replace.
     """
-    global grype_db_session
-
     # Get and release read locks
     read_lock = grype_db_lock.gen_rlock()
     if read_lock.acquire(blocking=False, timeout=60):
@@ -334,9 +424,13 @@ def query_vulnerabilities(
                 )
             )
 
-            query = grype_db_session.query(GrypeVulnerability).join(
-                GrypeVulnerabilityMetadata,
-                GrypeVulnerability.id == GrypeVulnerabilityMetadata.id,
+            query = (
+                _get_grype_db_session()
+                .query(GrypeVulnerability)
+                .join(
+                    GrypeVulnerabilityMetadata,
+                    GrypeVulnerability.id == GrypeVulnerabilityMetadata.id,
+                )
             )
 
             if vuln_id is not None:
